GPU routes file structure first draft
