+++ server/routes/components/gpu/index.js
@@ -1,0 +1,18 @@
+const router = require("express").Router();
+
+// Import routes here
+const gpuManufactured = require("./manufactured");
+const gpuReference = require("./reference");
+
+router
+  .get("/", (req, res) => {
+    console.log("API hit");
+    res.send({
+      express_says:
+        "This is the gpu section of the api. Please enter a gpu route for more information."
+    });
+  })
+  .use("/manufactured", gpuManufactured)
+  .use("/reference", gpuReference);
+
+module.exports = router;
